Moved job-factory to a proper place
